refactoring - core
